Avoid exceptions in mixed version cluster
1. Avoid following exceptions in mixed version clusters when new MQTT
   connections are created:
```
{{exception,{undef,[{rabbit_mqtt_util,remove_duplicate_clientid_connections,
                                      [{<<"/">>,
                                        <<"publish_to_all_queue_types">>},
                                       <0.1447.0>],
                                      []}]}},
 [{erpc,execute_cast,3,[{file,"erpc.erl"},{line,621}]}]}
```
If feature flag delete_ra_cluster_mqtt_node is disabled, let's still
populate pg with MQTT client IDs such that we don't have to migrate them
from the Ra cluster to pg when we enable the feature flag.
However, for actually closing duplicate MQTT client ID connections, if
that feature flag is disabled, let's rely on the Ra cluster to take care
of it.

2. Write a test ensuring the QoS responses are in the right order when a
   single SUBSCRIBE packet contains multiple subscriptions.